AGE-182: only write explicit changes to config file (#36)
The config system had a problem where the _effective_ config (after
resolving CLI args, env vars, and defaults) would be written to the
config file when _any_ config was intended to be saved.

This caused several problems, most annoying of which being that
temporarily passed CLI flags would get persisted (like `-o json`).

The solution is to never use the global `viper` instance to write the
config file. Instead, a new instance is created, and only populated with
the *file* content before being updated as needed. Then this can be
written without carrying in unwanted values.